Move shared GitHub client into common deploy helper
Committed-By-Agent: codex
